[69448] Send assignee notification email only once after PIR completion
https://community.openproject.org/wp/69448

In `WorkPackages::CreateService`, the work package was saved twice:
first time at creation, second time when rescheduling related work
packages, which includes self.

Each save created a `WorkflowJob`. Normally this is ok, but if there are
mentions in the description or journal_notes, and if the two jobs are
processed in parallel (in different GoodJob threads), this leads to
having the mention notification email sent twice. Indeed, each one will
try to create the notifications and send them. The `mail_alert_sent`
flag on the notification is not set to `true` until one mail has been
sent.

To fix it, the work packages that have not been changed after
rescheduling related work packages are filtered out. This guarantees
there is only one single save in the nominal cases where there are no
related work packages that would force a second save of the work
package. Single save -> single mention notification email.